First commit of exclude-hot-fix

+++ syncMyFolder.ps1
@@ -458,7 +458,13 @@ function Copy-Content {
     $updated_path = $target_folder + $path_split[1]
     if (!($fullPath -in $_source_exclusions)){
         Write-Log "{INFO}(Copy-Content) COPYING '$($fullPath)' IN FOLDER '$($updated_path)'"
-        Copy-Item -Path $fullPath -Destination $updated_path -Recurse               
+        if (!(Test-Path -Path $profile -PathType leaf)){
+            # If it's a folder, simply create it without its content
+            New-Item -Path $updated_path -ItemType Directory
+        } else {
+            # Else, if it's a file just copy to right destination
+            Copy-Item -Path $fullPath -Destination $updated_path -Recurse
+        }               
     } else {
         # If it is then log it
         Write-Log "{INFO}(Copy-Content) Exclude following file/folder from sync : '$($fullPath)'"
